# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Permissions:
    properties:
      product_module:
        type: boolean
      profile_module:
        type: boolean
      user_module:
        type: boolean
    type: object
...
  domain.Profile:
    properties:
      id:
        type: integer
      name:
        minLength: 4
        type: string
      permissions:
        $ref: '#/definitions/domain.Permissions'
    required:
    - name
    type: object
  domain.User:
    properties:
      id:
        type: integer
      mail:
        type: string
      name:
        minLength: 5
        type: string
      new:
        type: boolean
      profile:
        $ref: '#/definitions/domain.Profile'
      status:
        type: boolean
    required:
    - mail
    - name
    type: object
  dto.ItemsOutputDTO:
    properties:
      count:
        type: integer
      items: {}
    type: object
...
  dto.PasswordInputDTO:
    properties:
      password:
        example: secret
        type: string
      password_confirm:
        example: secret
        type: string
    type: object
...
  dto.UserInputDTO:
    properties:
      email:
        example: john.cena@email.com
        type: string
      name:
        example: John Cena
        type: string
      profile_id:
        example: 1
        type: integer
      status:
        example: true
        type: boolean
    type: object
  httphelper.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
info:
  contact:
    email: email@email.com
    name: Raul del Aguila
  description: Expenses API.
  title: Go - Expenses API
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Get all users
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - example: 10
        in: query
        name: limit
        type: integer
      - example: descending order 'desc' or ascending order 'asc'
        in: query
        name: order
        type: string
      - example: 1
        in: query
        name: page
        type: integer
      - example: 1
        in: query
        name: profile_id
        type: integer
      - example: name
        in: query
        name: search
        type: string
      - example: '''updated_at'', ''created_at'', ''name'' or some other field of
          the response object'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ItemsOutputDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Get users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Insert user
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: User model
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Insert user
      tags:
      - User
  /user/{email}/passw:
    patch:
      consumes:
      - application/json
      description: Set user password by ID
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: User email
        in: path
        name: email
        required: true
        type: string
      - description: Password model
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/dto.PasswordInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      summary: Set user password
      tags:
      - User
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user by ID
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Delete user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Get user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user by ID
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User model
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Update user
      tags:
      - User
  /user/{id}/reset:
    patch:
      consumes:
      - application/json
      description: Reset user password by ID
      parameters:
      - description: Language responses
        in: query
        name: lang
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httphelper.HTTPError'
      security:
      - Bearer: []
      summary: Reset user password
      tags:
      - User
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"